+++ ThreadsOfHistory/DataAccess/IRepository.cs
@@ -7,7 +7,7 @@ namespace DataAccess
 {
     public interface IRepository<T> : IDisposable where T : BaseModel
     {
-        Task<Event> GetAsync(decimal id);
+        Task<Event> GetAsync(long id);
         Task<List<Event>> GetAsync();
         void Add(T entity);
         void Update(T entity);
